requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dateparser==1.2.0
evdev==1.7.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==1.26.4
pandas==2.2.3
pandas-ta==0.3.14b0
propcache==0.2.0
pycryptodome==3.21.0
pynput==1.7.7
python-binance==1.0.24
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websockets==14.1
yarl==1.18.0
flask==3.0.0
flask-cors==4.0.0
pytest==8.0.0
scikit-learn==1.3.2
